# contracts/agent-integrator-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Platform delegated integrator API",
    "version": "1.0.0-preview.9",
    "description": "AP bearer reads, selected deployment discovery, delegated conversation replies and source refresh requests, action-bound authorized attachment bytes, signed Field reservation/facts event inbox, signed connection revoke, post-revoke event status recovery, and durable reservation notification route closure. Adds consent-bound installation-only public connections and origin-verified owned_embed preparation and management; native Field bidirectional bindings and selected conversation resources remain unchanged."
  },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "fieldEventHmac": {
        "type": "apiKey", "in": "header", "name": "X-Signature",
        "description": "HMAC-SHA256 over timestamp.event_id.raw_body_bytes using the connection key selected by X-Key-Id. X-Event-Id and X-Timestamp are required."
      },
      "fieldRevokeHmac": {
        "type": "apiKey", "in": "header", "name": "X-Signature",
        "description": "HMAC-SHA256 over timestamp.revocation_id.connection_id.revoke using the connection key selected by X-Key-Id. X-Revocation-Id and X-Timestamp are required. The same revocation ID can be retried."
      },
      "fieldRecoveryHmac": {
        "type": "apiKey", "in": "header", "name": "X-Signature",
        "description": "HMAC-SHA256 over timestamp.event_id.connection_id.notification-status using the connection key selected by X-Key-Id. X-Connection-Id and X-Timestamp are required. This read-only route is available only after AP connection revoke."
      },
      "fieldRouteCloseHmac": {
        "type": "apiKey", "in": "header", "name": "X-Signature",
        "description": "HMAC-SHA256 over timestamp.transferId.connectionId.actionRequestId.reservationId.latestRevision.latestEventId.route-close using the connection key selected by X-Key-Id. X-Timestamp is required. Only a revoked connection may close generation 1; identical transfer IDs may be retried."
      },
      "apOAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/auth/oauth2/authorize",
            "tokenUrl": "/api/auth/oauth2/token",
            "scopes": {
              "ap.agent.read": "Read the selected approved AI's minimal identity and revision",
              "ap.conversations.read": "Read consented conversations on explicitly selected active deployments",
              "ap.conversations.reply": "Write a human reply as the delegated AP owner",
              "ap.sources.refresh": "Request a review-only refresh of this connection's approved Field source",
              "ap.connections.create": "Create/read this client and consent grant installation-only binding; no Field data or booking access",
              "ap.deployments.manage": "Prepare, verify, activate and pause only owned_embed deployments created through this installation binding"
            }
          }
        }
      }
    },
    "schemas": {
...
      "FieldNotificationRouteClose": {
        "type": "object",
        "required": ["transferId", "connectionId", "actionRequestId", "reservationId", "latestRevision", "latestEventId", "routeGeneration"],
        "properties": {
          "transferId": { "type": "string", "format": "uuid" },
          "connectionId": { "type": "string", "format": "uuid" },
          "actionRequestId": { "type": "string", "format": "uuid" },
          "reservationId": { "type": "string", "format": "uuid" },
          "latestRevision": { "type": "integer", "minimum": 0 },
          "latestEventId": { "type": "string", "format": "uuid" },
          "routeGeneration": { "type": "integer", "enum": [2] }
        }, "additionalProperties": false
      },
      "FieldNotificationRouteClosure": {
        "allOf": [
          { "$ref": "#/components/schemas/FieldNotificationRouteClose" },
          { "type": "object", "required": ["closedAt"],
            "properties": { "closedAt": { "type": "string", "format": "date-time" } } }
        ]
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" }, "currentRevision": { "type": "integer", "minimum": 0 } },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Unauthorized": { "description": "Missing, expired, revoked, or invalid AP access token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "NotFound": { "description": "Resource outside this grant or not present", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  },
  "paths": {
...
    "/integrations/v1/notification-routes/close": {
      "post": {
        "operationId": "closeSignedFieldReservationNotificationRoute",
        "description": "After revoke, durably close AP generation 1 for one accepted Field reservation only when all events through the supplied revision are contiguous, processed, and customer notification states are resolved. The receipt is not a Field activation or a sent message.",
        "parameters": [
          { "in": "header", "name": "X-Key-Id", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "in": "header", "name": "X-Timestamp", "required": true, "schema": { "type": "string", "pattern": "^[0-9]{10}$" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FieldNotificationRouteClose" } } } },
        "security": [{ "fieldRouteCloseHmac": [] }],
        "responses": {
          "200": { "description": "Durable AP closure or identical transfer replay", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FieldNotificationRouteClosure" } } } },
          "400": { "description": "Malformed request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "description": "Missing, unresolved, or conflicting event or closure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Closure key unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}